LPS-116516 Create NewTopic component
